+++ platformio/__init__.py
@@ -34,4 +34,5 @@ __license__ = "Apache Software License"
 __copyright__ = "Copyright 2014-present PlatformIO"
 
 __apiurl__ = "https://api.platformio.org"
+__pioaccount_api__ = "https://api.accounts.platformio.org"
 __pioremote_endpoint__ = "ssl:remote.platformio.org:4413"